+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fixed AutoEquatable access level for `==` func
 - Fixed path of generated files when linked to Xcode project
+- Fixed  extraction of inline annotations in comments like /** sourcery: skipDescription */
 
 ## 0.13.1