+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ---
 
-## Master
+## 0.14.0
 
 ### New Features